requirements.txt
# Multi-Agent Framework
crewai

# LLM Orchestration & Tools
crewai_tools
pydantic

# UI and Dashboard
streamlit

# Data Retrieval and Environment
yfinance
python-dotenv